# packages/shared_schemas/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "ConstantAttenuationResult": {
        "additionalProperties": false,
        "properties": {
          "attenuation_db_per_km": {
            "minimum": 0,
            "title": "Attenuation Db Per Km",
            "type": "number"
          },
          "distance_samples_km": {
            "items": {
              "type": "number"
            },
            "maxItems": 65,
            "minItems": 1,
            "title": "Distance Samples Km",
            "type": "array"
          },
          "input_power_dbm": {
            "title": "Input Power Dbm",
            "type": "number"
          },
          "length_km": {
            "minimum": 0,
            "title": "Length Km",
            "type": "number"
          },
          "model_manifest": {
            "$ref": "#/components/schemas/ConstantAttenuationManifest"
          },
          "output_power_dbm": {
            "title": "Output Power Dbm",
            "type": "number"
          },
          "power_samples_dbm": {
            "items": {
              "type": "number"
            },
            "maxItems": 65,
            "minItems": 1,
            "title": "Power Samples Dbm",
            "type": "array"
          },
          "section_loss_db": {
            "minimum": 0,
            "title": "Section Loss Db",
            "type": "number"
          }
        },
        "required": [
          "length_km",
          "attenuation_db_per_km",
          "input_power_dbm",
          "section_loss_db",
          "output_power_dbm",
          "distance_samples_km",
          "power_samples_dbm",
          "model_manifest"
        ],
        "title": "ConstantAttenuationResult",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GaussianModeProfileResult": {
        "additionalProperties": false,
        "properties": {
          "grid_half_width_um": {
            "exclusiveMinimum": 0,
            "title": "Grid Half Width Um",
            "type": "number"
          },
          "grid_points": {
            "maximum": 65,
            "minimum": 3,
            "title": "Grid Points",
            "type": "integer"
          },
          "mode_field_radius_um": {
            "exclusiveMinimum": 0,
            "title": "Mode Field Radius Um",
            "type": "number"
          },
          "model_manifest": {
            "$ref": "#/components/schemas/GaussianModeProfileManifest"
          },
          "normalized_field": {
            "items": {
              "items": {
                "maximum": 1,
                "minimum": 0,
                "type": "number"
              },
              "type": "array"
            },
            "title": "Normalized Field",
            "type": "array"
          },
          "normalized_intensity": {
            "items": {
              "items": {
                "maximum": 1,
                "minimum": 0,
                "type": "number"
              },
              "type": "array"
            },
            "title": "Normalized Intensity",
            "type": "array"
          },
          "x_um": {
            "items": {
              "type": "number"
            },
            "title": "X Um",
            "type": "array"
          },
          "y_um": {
            "items": {
              "type": "number"
            },
            "title": "Y Um",
            "type": "array"
          }
        },
        "required": [
          "mode_field_radius_um",
          "grid_half_width_um",
          "grid_points",
          "x_um",
          "y_um",
          "normalized_field",
          "normalized_intensity",
          "model_manifest"
        ],
        "title": "GaussianModeProfileResult",
        "type": "object"
      },
...
      "GroupDelayResult": {
        "additionalProperties": false,
        "properties": {
          "group_delay_ps": {
            "minimum": 0,
            "title": "Group Delay Ps",
            "type": "number"
          },
          "group_index_dimensionless": {
            "exclusiveMinimum": 0,
            "title": "Group Index Dimensionless",
            "type": "number"
          },
          "length_km": {
            "minimum": 0,
            "title": "Length Km",
            "type": "number"
          },
          "model_manifest": {
            "$ref": "#/components/schemas/GroupDelayManifest"
          }
        },
        "required": [
          "length_km",
          "group_index_dimensionless",
          "group_delay_ps",
          "model_manifest"
        ],
        "title": "GroupDelayResult",
        "type": "object"
      },
      "GuidanceModelManifest": {
        "additionalProperties": false,
        "properties": {
          "assumptions": {
            "default": [
              "ideal circular step-index profile",
              "scalar weak-guidance mode interpretation",
              "homogeneous, isotropic, linear media",
              "n_external=1 for air angle"
            ],
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "limitations": {
            "default": [
              "asymptotic mode count only at V >= 10.0 project threshold",
              "V=2.405 ideal cutoff distinct from measured cable cutoff",
              "not a G.652.D conformance model"
            ],
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "mode_count_min_v_dimensionless": {
            "default": 10.0,
            "title": "Mode Count Min V Dimensionless",
            "type": "number"
          },
          "mode_regime_cutoff_v_dimensionless": {
            "default": 2.405,
            "title": "Mode Regime Cutoff V Dimensionless",
            "type": "number"
          },
          "model_id": {
            "const": "ideal_circular_step_index_guidance",
            "default": "ideal_circular_step_index_guidance",
            "title": "Model Id",
            "type": "string"
          },
          "model_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Model Version",
            "type": "string"
          }
        },
        "title": "GuidanceModelManifest",
        "type": "object"
      },
      "GuidanceRequest": {
        "additionalProperties": false,
        "properties": {
          "core_radius_um": {
            "exclusiveMinimum": 0,
            "title": "Core Radius Um",
            "type": "number"
          },
          "n_cladding": {
            "exclusiveMinimum": 0,
            "title": "N Cladding",
            "type": "number"
          },
          "n_core": {
            "exclusiveMinimum": 0,
            "title": "N Core",
            "type": "number"
          },
          "wavelength_nm": {
            "exclusiveMinimum": 0,
            "title": "Wavelength Nm",
            "type": "number"
          }
        },
        "required": [
          "n_core",
          "n_cladding",
          "core_radius_um",
          "wavelength_nm"
        ],
        "title": "GuidanceRequest",
        "type": "object"
      },
      "GuidanceResult": {
        "additionalProperties": false,
        "properties": {
          "air_acceptance_angle_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Air Acceptance Angle Deg"
          },
          "approximate_mode_count": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Approximate Mode Count"
          },
          "critical_angle_deg": {
            "title": "Critical Angle Deg",
            "type": "number"
          },
          "mode_regime": {
            "$ref": "#/components/schemas/ModeRegime"
          },
          "model_manifest": {
            "$ref": "#/components/schemas/GuidanceModelManifest"
          },
          "numerical_aperture_dimensionless": {
            "title": "Numerical Aperture Dimensionless",
            "type": "number"
          },
          "relative_index_difference_dimensionless": {
            "title": "Relative Index Difference Dimensionless",
            "type": "number"
          },
          "v_number_dimensionless": {
            "title": "V Number Dimensionless",
            "type": "number"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/GuidanceWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "critical_angle_deg",
          "numerical_aperture_dimensionless",
          "relative_index_difference_dimensionless",
          "v_number_dimensionless",
          "mode_regime",
          "warnings",
          "model_manifest"
        ],
        "title": "GuidanceResult",
        "type": "object"
      },
      "GuidanceWarning": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/GuidanceWarningCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "output_field": {
            "enum": [
              "air_acceptance_angle_deg",
              "approximate_mode_count"
            ],
            "title": "Output Field",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "output_field"
        ],
        "title": "GuidanceWarning",
        "type": "object"
      },
...
      "Level1FibreConfig": {
        "additionalProperties": false,
        "properties": {
          "attenuation_db_per_km": {
            "minimum": 0,
            "title": "Attenuation Db Per Km",
            "type": "number"
          },
          "cable_application": {
            "$ref": "#/components/schemas/G652DAttenuationApplication"
          },
          "core_radius_um": {
            "exclusiveMinimum": 0,
            "title": "Core Radius Um",
            "type": "number"
          },
          "dispersion_ps_per_nm_km": {
            "title": "Dispersion Ps Per Nm Km",
            "type": "number"
          },
          "group_index_dimensionless": {
            "exclusiveMinimum": 0,
            "title": "Group Index Dimensionless",
            "type": "number"
          },
          "mode_field_radius_um": {
            "exclusiveMinimum": 0,
            "title": "Mode Field Radius Um",
            "type": "number"
          },
          "n_cladding": {
            "exclusiveMinimum": 0,
            "title": "N Cladding",
            "type": "number"
          },
          "n_core": {
            "exclusiveMinimum": 0,
            "title": "N Core",
            "type": "number"
          }
        },
        "required": [
          "n_core",
          "n_cladding",
          "core_radius_um",
          "mode_field_radius_um",
          "attenuation_db_per_km",
          "dispersion_ps_per_nm_km",
          "group_index_dimensionless",
          "cable_application"
        ],
        "title": "Level1FibreConfig",
        "type": "object"
      },
      "Level1FibrePreset": {
        "enum": [
          "custom",
          "g652d"
        ],
        "title": "Level1FibrePreset",
        "type": "string"
      },
...
      "Level1SamplingConfig": {
        "additionalProperties": false,
        "properties": {
          "grid_half_width_um": {
            "exclusiveMinimum": 0,
            "title": "Grid Half Width Um",
            "type": "number"
          },
          "grid_points": {
            "default": 65,
            "maximum": 65,
            "minimum": 3,
            "title": "Grid Points",
            "type": "integer"
          }
        },
        "required": [
          "grid_half_width_um"
        ],
        "title": "Level1SamplingConfig",
        "type": "object"
      },
      "Level1SectionConfig": {
        "additionalProperties": false,
        "properties": {
          "length_km": {
            "minimum": 0,
            "title": "Length Km",
            "type": "number"
          }
        },
        "required": [
          "length_km"
        ],
        "title": "Level1SectionConfig",
        "type": "object"
      },
      "Level1SimulationManifest": {
        "additionalProperties": false,
        "properties": {
          "assumptions": {
            "default": [
              "one uniform fibre section",
              "all calculations share one operating wavelength",
              "fibre composition is uniform over the section"
            ],
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "component_model_ids": {
            "items": {
              "type": "string"
            },
            "title": "Component Model Ids",
            "type": "array"
          },
          "limitations": {
            "default": [
              "excludes bends, splices, and connectors",
              "excludes polarization-mode dispersion",
              "excludes optical nonlinearity",
              "excludes multi-section links",
              "excludes full-wave field solving"
            ],
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "model_id": {
            "const": "level1_single_section_simulation",
            "default": "level1_single_section_simulation",
            "title": "Model Id",
            "type": "string"
          },
          "model_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Model Version",
            "type": "string"
          }
        },
        "required": [
          "component_model_ids"
        ],
        "title": "Level1SimulationManifest",
        "type": "object"
      },
      "Level1SimulationRequest": {
        "additionalProperties": false,
        "properties": {
          "fibre": {
            "$ref": "#/components/schemas/Level1FibreConfig"
          },
          "preset": {
            "$ref": "#/components/schemas/Level1FibrePreset"
          },
          "sampling": {
            "$ref": "#/components/schemas/Level1SamplingConfig"
          },
          "section": {
            "$ref": "#/components/schemas/Level1SectionConfig"
          },
          "source": {
            "$ref": "#/components/schemas/Level1SourceConfig"
          }
        },
        "required": [
          "preset",
          "fibre",
          "source",
          "section",
          "sampling"
        ],
        "title": "Level1SimulationRequest",
        "type": "object"
      },
      "Level1SimulationResult": {
        "additionalProperties": false,
        "properties": {
          "attenuation": {
            "$ref": "#/components/schemas/ConstantAttenuationResult"
          },
          "configuration": {
            "$ref": "#/components/schemas/Level1SimulationRequest"
          },
          "group_delay": {
            "$ref": "#/components/schemas/GroupDelayResult"
          },
          "guidance": {
            "$ref": "#/components/schemas/GuidanceResult"
          },
          "mode_profile": {
            "$ref": "#/components/schemas/GaussianModeProfileResult"
          },
          "model_manifest": {
            "$ref": "#/components/schemas/Level1SimulationManifest"
          },
          "parameter_boundaries": {
            "items": {
              "$ref": "#/components/schemas/Level1ParameterBoundary"
            },
            "title": "Parameter Boundaries",
            "type": "array"
          },
          "pulse_broadening": {
            "$ref": "#/components/schemas/ChromaticPulseBroadeningResult"
          },
          "standards_checks": {
            "$ref": "#/components/schemas/Level1StandardsChecks"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Level1Warning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "configuration",
          "guidance",
          "mode_profile",
          "attenuation",
          "group_delay",
          "pulse_broadening",
          "standards_checks",
          "parameter_boundaries",
          "warnings",
          "model_manifest"
        ],
        "title": "Level1SimulationResult",
        "type": "object"
      },
      "Level1SourceConfig": {
        "additionalProperties": false,
        "properties": {
          "input_power_dbm": {
            "title": "Input Power Dbm",
            "type": "number"
          },
          "input_pulse_fwhm_ps": {
            "exclusiveMinimum": 0,
            "title": "Input Pulse Fwhm Ps",
            "type": "number"
          },
          "spectral_width_fwhm_nm": {
            "minimum": 0,
            "title": "Spectral Width Fwhm Nm",
            "type": "number"
          },
          "wavelength_nm": {
            "exclusiveMinimum": 0,
            "title": "Wavelength Nm",
            "type": "number"
          }
        },
        "required": [
          "wavelength_nm",
          "input_power_dbm",
          "spectral_width_fwhm_nm",
          "input_pulse_fwhm_ps"
        ],
        "title": "Level1SourceConfig",
        "type": "object"
      },
...
      "Level1SweepRequest": {
        "additionalProperties": false,
        "properties": {
          "base_configuration": {
            "$ref": "#/components/schemas/Level1SimulationRequest"
          },
          "parameter": {
            "$ref": "#/components/schemas/Level1SweepParameter"
          },
          "sample_count": {
            "description": "Number of independently evaluated parameter values.",
            "maximum": 200,
            "minimum": 2,
            "title": "Sample Count",
            "type": "integer"
          },
          "start_value": {
            "description": "First parameter value, using the unit implied by parameter.",
            "title": "Start Value",
            "type": "number"
          },
          "stop_value": {
            "description": "Last parameter value, using the unit implied by parameter.",
            "title": "Stop Value",
            "type": "number"
          }
        },
        "required": [
          "base_configuration",
          "parameter",
          "start_value",
          "stop_value",
          "sample_count"
        ],
        "title": "Level1SweepRequest",
        "type": "object"
      },
      "Level1SweepResult": {
        "additionalProperties": false,
        "properties": {
          "model_manifest": {
            "$ref": "#/components/schemas/Level1SweepManifest"
          },
          "parameter_unit": {
            "description": "Unit shared by request start_value, stop_value, and point parameter_value.",
            "enum": [
              "dimensionless",
              "\u00b5m",
              "dB/km",
              "ps/(nm\u00b7km)",
              "nm",
              "dBm",
              "ps",
              "km"
            ],
            "title": "Parameter Unit",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/Level1SweepPoint"
            },
            "maxItems": 200,
            "minItems": 2,
            "title": "Points",
            "type": "array"
          },
          "request": {
            "$ref": "#/components/schemas/Level1SweepRequest"
          }
        },
        "required": [
          "request",
          "parameter_unit",
          "points",
          "model_manifest"
        ],
        "title": "Level1SweepResult",
        "type": "object"
      },
      "Level1Warning": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/Level1WarningCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "output_field": {
            "title": "Output Field",
            "type": "string"
          },
          "source_model_id": {
            "title": "Source Model Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "source_model_id",
          "message",
          "output_field"
        ],
        "title": "Level1Warning",
        "type": "object"
      },
...
      "ModeRegime": {
        "enum": [
          "single_mode",
          "multimode"
        ],
        "title": "ModeRegime",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Optical Fibre Simulator API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/guidance/calculate": {
      "post": {
        "operationId": "calculate_guidance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Post Guidance Calculate"
      }
    },
...
    "/api/v1/simulations/preview": {
      "post": {
        "operationId": "preview_level1_simulation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Level1SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level1SimulationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation or calculation failed"
          }
        },
        "summary": "Post Simulation Preview"
      }
    },
    "/api/v1/simulations/sweep": {
      "post": {
        "operationId": "sweep_level1_parameter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Level1SweepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level1SweepResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation or calculation failed"
          }
        },
        "summary": "Post Simulation Sweep"
      }
    }
  }
}